Also show help if `sst` is given no arguments at all.
I also made the help output to stdout/stderr depending on if the help
appears alone or not used at all (--> likely sst is doing the intended
thing by printing help, so stdout and exit code 0, and conversely if
--help is among other options, output to stderr and exit code 1).

I wonder if I have to consider command-line arguments to be a threat
vector. I.e. untrusted input slips into sst command-line, can it be
meaningfully defended against? Should you? This change is slightly
motivated by the idea that -h or --help slips *accidentally* into the
command line; if in a pipeline we output help to stdout then we might
mess up a downstream program. Outputting to stderr seems safer.

I've not really written programs that got so popular anywhere close that
I've had to personally see how this stuff can fail or cause havoc.