Drop Landlock ABI 4 support.
This is because `sst` now actively uses a feature introduced in ABI 5
(ioctl dev).

I may check later how old is ABI 4 exactly. If it is from 2005 then I am
unlikely to add this back. If it is from 2023, then I likely will
investigate if I need to do anything to add it back.